movies/requirements.txt
django==5.0.4
djangorestframework==3.15.1
PyJWT==2.8.0
requests==2.31.0
python-dotenv
factory_boy